refactor: extract the invitation service and split admin routes into modules
- domain/invitations.ts: inviteMember()/resendInvitation() own token
  creation, the 7-day expiry, the record fetch and email delivery — the
  block that was copy-pasted across invite, create-member and resend.
- routes/admin/{index,settings,providers,invitations,members,audit}.ts
  replace the 800-line admin.ts; the index mounts the sub-routers behind the
  household + owner middleware. Behaviour and paths are unchanged.

Closes #110
